Move unparameterized grammars to new unparam package
This leaves parameterized grammars alone in syntax. The new package will
contain analysis and monomorphization (aka expansion) of parameterized
grammars.
