+++ Dolmen/src/jg/JGParserGenerated.java
@@ -6,10 +6,10 @@ import java.util.ArrayList;
 import common.Lists;
 import syntax.Extent;
 import syntax.Located;
-import syntax.Production;
+import unparam.Production;
 import syntax.TokenDecl;
-import syntax.GrammarRule;
-import syntax.Grammar;
+import unparam.GrammarRule;
+import unparam.Grammar;
 
 /**
  * Parser generated by Dolmen 1.0.0
