+++ Dolmen/src/test/examples/JSonParser.java
@@ -10,7 +10,7 @@ import common.Maps;
 /**
  * Parser generated by Dolmen 1.0.0
  */
-public final class JSonParser extends codegen.BaseParser<JSonParser.Token> {
+public final class JSonParser extends codegen.BaseParser.WithPositions<JSonParser.Token> {
     
     @SuppressWarnings("javadoc")
     public static abstract class Token {
@@ -261,48 +261,65 @@ public final class JSonParser extends codegen.BaseParser<JSonParser.Token> {
      * Entry point for the non-terminal json
      */
     public  Value<?>  json() {
+        enter(2);
         // v = value
          Value<?>  v = value();
+        leave("v");
         // EOF
         eat(Token.Kind.EOF);
+        shift(null);
          return v; 
     }
     
     private  Value<?>  value() {
         switch (peek().getKind()) {
             case FALSE: {
+                enter(1);
                 // FALSE
                 eat(Token.Kind.FALSE);
+                shift(null);
                  return valFalse; 
             }
             case LBRACKET: {
+                enter(1);
                 // o = object
                  Map<String, Value<?>>  o = object();
+                leave("o");
                  return valObject(o); 
             }
             case LSQUARE: {
+                enter(1);
                 // a = array
                  List<Value<?>>  a = array();
+                leave("a");
                  return valArray(a); 
             }
             case NULL: {
+                enter(1);
                 // NULL
                 eat(Token.Kind.NULL);
+                shift(null);
                  return valNull; 
             }
             case NUMBER: {
+                enter(1);
                 // n = NUMBER
                  double  n = ((Token.NUMBER) eat(Token.Kind.NUMBER)).value;
+                shift("n");
                  return valNumber(n); 
             }
             case STRING: {
+                enter(1);
                 // s = STRING
                  String  s = ((Token.STRING) eat(Token.Kind.STRING)).value;
+                shift("s");
                  return valString(s); 
             }
             case TRUE: {
+                enter(1);
                 // TRUE
                 eat(Token.Kind.TRUE);
+                shift(null);
                  return valTrue; 
             }
             default: {
@@ -312,10 +329,13 @@ public final class JSonParser extends codegen.BaseParser<JSonParser.Token> {
     }
     
     private  List<Value<?>>  array() {
+        enter(2);
         // LSQUARE
         eat(Token.Kind.LSQUARE);
+        shift(null);
         // elts = elements(null)
          List<Value<?>>  elts = elements(null);
+        leave("elts");
          return elts; 
     }
     
@@ -328,17 +348,22 @@ public final class JSonParser extends codegen.BaseParser<JSonParser.Token> {
             case NUMBER:
             case STRING:
             case TRUE: {
+                enter(2);
                 // val = value
                  Value<?>  val = value();
+                leave("val");
                  List<Value<?>> acc = elts == null ? new ArrayList<>() : elts; 
                  acc.add(val); 
                 // more_elements(acc)
                 more_elements(acc);
+                leave(null);
                  return acc; 
             }
             case RSQUARE: {
+                enter(1);
                 // RSQUARE
                 eat(Token.Kind.RSQUARE);
+                shift(null);
                  return elts == null ? Lists.empty() : elts; 
             }
             default: {
@@ -348,20 +373,25 @@ public final class JSonParser extends codegen.BaseParser<JSonParser.Token> {
     }
     
     private  void  more_elements(List<Value<?>> elts) {
+        enter(2);
         more_elements:
         while (true) {
             switch (peek().getKind()) {
                 case COMMA: {
                     // COMMA
                     eat(Token.Kind.COMMA);
+                    shift(null);
                     // val = value
                      Value<?>  val = value();
+                    leave("val");
                      elts.add(val); 
+                    rewind();
                     continue more_elements;
                 }
                 case RSQUARE: {
                     // RSQUARE
                     eat(Token.Kind.RSQUARE);
+                    shift(null);
                      return; 
                 }
                 default: {
@@ -372,26 +402,34 @@ public final class JSonParser extends codegen.BaseParser<JSonParser.Token> {
     }
     
     private  Map<String, Value<?>>  object() {
+        enter(2);
         // LBRACKET
         eat(Token.Kind.LBRACKET);
+        shift(null);
         // members = members(null)
          Map<String, Value<?>>  members = members(null);
+        leave("members");
          return members; 
     }
     
     private  Map<String, Value<?>>  members(@Nullable Map<String, Value<?>> members) {
         switch (peek().getKind()) {
             case RBRACKET: {
+                enter(1);
                 // RBRACKET
                 eat(Token.Kind.RBRACKET);
+                shift(null);
                  return members == null ? Maps.empty() : members; 
             }
             case STRING: {
+                enter(2);
                  Map<String, Value<?>> acc = members == null ? new HashMap<>() : members; 
                 // pair(acc)
                 pair(acc);
+                leave(null);
                 // more_members(acc)
                 more_members(acc);
+                leave(null);
                  return acc; 
             }
             default: {
@@ -401,19 +439,24 @@ public final class JSonParser extends codegen.BaseParser<JSonParser.Token> {
     }
     
     private  void  more_members(Map<String, Value<?>> members) {
+        enter(2);
         more_members:
         while (true) {
             switch (peek().getKind()) {
                 case COMMA: {
                     // COMMA
                     eat(Token.Kind.COMMA);
+                    shift(null);
                     // pair(members)
                     pair(members);
+                    leave(null);
+                    rewind();
                     continue more_members;
                 }
                 case RBRACKET: {
                     // RBRACKET
                     eat(Token.Kind.RBRACKET);
+                    shift(null);
                      return; 
                 }
                 default: {
@@ -424,12 +467,16 @@ public final class JSonParser extends codegen.BaseParser<JSonParser.Token> {
     }
     
     private  void  pair(Map<String, Value<?>> map) {
+        enter(3);
         // s = STRING
          String  s = ((Token.STRING) eat(Token.Kind.STRING)).value;
+        shift("s");
         // COLON
         eat(Token.Kind.COLON);
+        shift(null);
         // val = value
          Value<?>  val = value();
+        leave("val");
          map.put(s, val); return; 
     }
